chore: Support floating and integer waypoints.
